# controllers/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "go-auction",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auction/new": {
            "post": {
                "description": "New auction for product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "Start new auction",
                "operationId": "new-auction",
                "parameters": [
                    {
                        "description": "New Auction Data",
                        "name": "userInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewAuction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auction succesfully created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auction/{id}": {
            "get": {
                "description": "Get a auction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "Get Auction",
                "operationId": "get-auction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get Auction Data",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succesful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Login For User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "operationId": "auth-login",
                "parameters": [
                    {
                        "description": "User login data",
                        "name": "userInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succesful register",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "User Logout Funciton",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout User",
                "operationId": "logout-user",
                "responses": {
                    "200": {
                        "description": "Succesful register",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register For New User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register For New User",
                "operationId": "auth-register",
                "parameters": [
                    {
                        "description": "User register data",
                        "name": "userInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succesful register",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/update": {
            "patch": {
                "description": "Update Infos For User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Update User",
                "operationId": "update-user",
                "parameters": [
                    {
                        "description": "User update data",
                        "name": "userInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated succesfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/offers/new": {
            "post": {
                "description": "Post new bid for auction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "New Offer For Auction",
                "operationId": "new-offer-auction",
                "parameters": [
                    {
                        "description": "New Offer Info",
                        "name": "newOfferInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OfferInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Offer sent",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/offers/{id}": {
            "get": {
                "description": "Get Real-Time Bids for auction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "Get Auction",
                "operationId": "get-ws-auction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get Auction",
                        "name": "auctionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Offer got",
                        "schema": {
                            "$ref": "#/definitions/models.OfferInfo"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/product/delete": {
            "delete": {
                "description": "Delete a product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Delete Product",
                "operationId": "delete-product",
                "parameters": [
                    {
                        "description": "Delete a product",
                        "name": "productID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DeleteProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted succesfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/product/new": {
            "post": {
                "description": "Add a product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "New Product",
                "operationId": "new-product",
                "parameters": [
                    {
                        "description": "New product data",
                        "name": "newProductInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewProductInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product created succesfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/product/update": {
            "patch": {
                "description": "Update a product infos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Update Product",
                "operationId": "update-product",
                "parameters": [
                    {
                        "description": "Update product data",
                        "name": "getProductInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GetProductInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product updated succesfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/product/{slug}": {
            "get": {
                "description": "Get Product Details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Get Product",
                "operationId": "get-product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Get succesfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.DeleteProduct": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                }
            }
        },
        "models.GetProductInfo": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "brand": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "is_open_to_sell": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "seller_id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "models.LoginModel": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.NewAuction": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "seller_id": {
                    "type": "string"
                },
                "start_price": {
                    "type": "number"
                }
            }
        },
        "models.NewProductInfo": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "is_open_to_sell": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "seller_id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "models.OfferInfo": {
            "type": "object",
            "properties": {
                "auction_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "offer": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.RegisterModel": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}